expand wording about database creation parameters
I’m now more convinced than ever that setting the DB-wide collation
to 'C' in the default case is the only correct thing to do. The
encoding can be SQL_ASCII (for speed) or (if you actually use the
changes in the database, i.e. it’s not just a backing store for a
higher-level layer or ORM) UTF-8 (for correctness) still.